+++ mobile/lib/common/settings/share_logs_screen.dart
@@ -2,20 +2,29 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:get_10101/common/application/switch.dart';
 import 'package:get_10101/common/color.dart';
 import 'package:get_10101/common/settings/settings_screen.dart';
 import 'package:get_10101/logger/hybrid_logger.dart';
 import 'package:get_10101/logger/logger.dart';
+import 'package:get_10101/util/preferences.dart';
 import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:share_plus/share_plus.dart';
 
-class ShareLogsScreen extends StatelessWidget {
+class ShareLogsScreen extends StatefulWidget {
   static const route = "${SettingsScreen.route}/$subRouteName";
   static const subRouteName = "sharelogs";
 
   const ShareLogsScreen({super.key});
+
+  @override
+  State<ShareLogsScreen> createState() => _ShareLogsScreenState();
+}
+
+class _ShareLogsScreenState extends State<ShareLogsScreen> {
+  bool changedLogLevel = false;
 
   @override
   Widget build(BuildContext context) {
@@ -113,6 +122,63 @@ class ShareLogsScreen extends StatelessWidget {
                           ],
                         ),
                       ),
+                    ),
+                    const SizedBox(height: 20),
+                    Container(
+                      padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+                      decoration: BoxDecoration(
+                          color: Colors.white, borderRadius: BorderRadius.circular(15)),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: [
+                          Text(
+                            "Set log level to trace",
+                            style: TextStyle(
+                                color: tenTenOnePurple.shade800,
+                                fontSize: 16,
+                                fontWeight: FontWeight.w500),
+                          ),
+                          FutureBuilder(
+                              future: Preferences.instance.isLogLevelTrace(),
+                              builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
+                                if (!snapshot.hasData) {
+                                  return Container();
+                                }
+
+                                logger.i("trace: ${snapshot.data}");
+                                return TenTenOneSwitch(
+                                    value: snapshot.data ?? false,
+                                    onChanged: (value) {
+                                      setState(() {
+                                        Preferences.instance.setLogLevelTrace(value);
+                                        changedLogLevel = true;
+                                      });
+                                    });
+                              }),
+                        ],
+                      ),
+                    ),
+                    const SizedBox(height: 20),
+                    Visibility(
+                      visible: changedLogLevel,
+                      child: Container(
+                          padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+                          decoration: BoxDecoration(
+                              border: Border.all(color: tenTenOnePurple),
+                              color: Colors.white,
+                              borderRadius: BorderRadius.circular(10)),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.start,
+                            children: [
+                              Icon(
+                                Icons.info_outline,
+                                color: tenTenOnePurple.shade800,
+                                size: 22,
+                              ),
+                              const SizedBox(width: 10),
+                              const Text("Restart the app to apply your changes."),
+                            ],
+                          )),
                     )
                   ],
                 ),