+++ mobile/lib/util/preferences.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:get_10101/features/stable/stable_screen.dart';
 import 'package:get_10101/features/trade/trade_screen.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -10,6 +11,17 @@ class Preferences {
   static const emailAddress = "emailAddress";
   static const openPosition = "openPosition";
   static const fullBackup = "fullBackup";
+  static const logLevelTrace = "logLevelTrace";
+
+  Future<bool> setLogLevelTrace(bool trace) async {
+    SharedPreferences preferences = await SharedPreferences.getInstance();
+    return preferences.setBool(logLevelTrace, trace);
+  }
+
+  Future<bool> isLogLevelTrace() async {
+    SharedPreferences preferences = await SharedPreferences.getInstance();
+    return preferences.getBool(logLevelTrace) ?? kDebugMode;
+  }
 
   Future<bool> setFullBackupRequired(bool required) async {
     SharedPreferences preferences = await SharedPreferences.getInstance();