+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add support for Android 9
+- Add settings option to enable trace logs
 
 ## [1.6.4] - 2023-11-24